Changes are now made in matlab
It seems that the only option is now to make the changes in matlab and to run this code mid analysis code. It currently still reads off of the .EVT files, even though the information is actually read in through the .HDR files. The EVT is a simplified format of part of the HDR file.